requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pillow
reportlab